Rename SynapseBackup to BackupObserver and split logic in two files
