+++ src-docs/backup_observer.py.md
@@ -1,0 +1,45 @@
+<!-- markdownlint-disable -->
+
+<a href="../src/backup_observer.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `backup_observer.py`
+S3 Backup relation observer for Synapse. 
+
+**Global Variables**
+---------------
+- **S3_CANNOT_ACCESS_BUCKET**
+- **S3_INVALID_CONFIGURATION**
+- **BACKUP_STATUS_MESSAGES**
+
+
+---
+
+## <kbd>class</kbd> `BackupObserver`
+The S3 backup relation observer. 
+
+<a href="../src/backup_observer.py#L26"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+### <kbd>function</kbd> `__init__`
+
+```python
+__init__(charm: CharmBase)
+```
+
+Initialize the backup object. 
+
+
+
+**Args:**
+ 
+ - <b>`charm`</b>:  The parent charm the backups are made for. 
+
+
+---
+
+#### <kbd>property</kbd> model
+
+Shortcut for more simple access the model. 
+
+
+
+
